% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2.R
\name{mapViridis}
\alias{mapViridis}
\title{Map vector of numeric values to Viridis color scale.}
\usage{
mapViridis(values, domain_min = 0, domain_max = 50, n = 256)
}
\arguments{
\item{values}{vector of numeric values (mostly -log10(p-values))}

\item{domain_min}{numeric value that corresponds to the 'yellow' in the color scale}

\item{domain_max}{numeric value that corresponds to the 'dark blue' in the color scale}

\item{n}{number of bins to generate from the color scale}
}
\value{
The output is a corresponding vector of colors from the Viridis color scale with domain in range(domain_min, domain_max).
}
\description{
Map vector of numeric values to Viridis color scale.
}
\author{
Liis Kolberg <liis.kolberg@ut.ee>
}
