% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2.R
\name{random_query}
\alias{random_query}
\title{Generate a random gene list for testing.}
\usage{
random_query(organism = "hsapiens")
}
\arguments{
\item{organism}{organism name. Organism names are constructed by concatenating the first letter of the name and the
family name. Example: human - 'hsapiens', mouse - 'mmusculus'.}
}
\value{
a character vector containing randomly selected gene IDs from the selected organism.
}
\description{
This function returns a vector of randomly selected genes from the selected organism.
}
\examples{
random_genes <- random_query()
}
\author{
Liis Kolberg <liis.kolberg@ut.ee>
}
