% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2.R
\name{set_tls_version}
\alias{set_tls_version}
\title{Set the TLS version to use for SSL}
\usage{
set_tls_version(v)
}
\arguments{
\item{v}{version: "1.2" (default), "1.1" (fallback)}
}
\description{
Set the TLS version. Could be useful at environments where SSL was built without TLS 1.2 support
}
