% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2.R
\name{set_user_agent}
\alias{set_user_agent}
\title{Set custom user agent string.}
\usage{
set_user_agent(ua, append = F)
}
\arguments{
\item{ua}{the user agent string.}

\item{append}{logical indicating whether to append the passed string to the default user agent string.}
}
\description{
Set the HTTP User-Agent string. Useful for overriding the default user agent for
packages that depend on gprofiler2 functionality.
}
