% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler2.R
\name{set_base_url}
\alias{set_base_url}
\title{Set the base URL.}
\usage{
set_base_url(url)
}
\arguments{
\item{url}{the base URL.}
}
\description{
Function to change the g:Profiler base URL. Useful for overriding the default URL (https://biit.cs.ut.ee/gprofiler)
with the beta (https://biit.cs.ut.ee/gprofiler_beta) or an archived version (available starting from the version e94_eg41_p11).
}
